blog writing agent/requirements.txt
# ── Core LLM & Agent Framework ─────────────────────────────────────────────
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.2.0

# ── Google AI ───────────────────────────────────────────────────────────────
google-genai>=0.7.0

# ── Research ────────────────────────────────────────────────────────────────
tavily-python>=0.3.0

# ── Data validation ─────────────────────────────────────────────────────────
pydantic>=2.0.0

# ── Environment ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Frontend ────────────────────────────────────────────────────────────────
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.20.0

# ── Database (PostgreSQL + pgvector) ────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# ── Resilience ──────────────────────────────────────────────────────────────
tenacity>=8.2.0

# ── Markdown → HTML export ──────────────────────────────────────────────────
markdown>=3.5.0
Pygments>=2.17.0

# ── Testing ─────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
